requirements.txt
flask
flask-sqlalchemy
sqlalchemy
python-dotenv
flask-smorest
flask-jwt-extended
passlib
gunicorn
flask_migrate
